# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: HideAndSeek
  description: Geographic Hide and Seek game server
  version: 0.1.0
paths:
...
  /games:
    post:
      tags:
      - games
      summary: Create Game
      description: Create a new game on a map.
      operationId: create_game_games_post
      parameters:
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/join:
    post:
      tags:
      - games
      summary: Join Game
      description: Join a game by its join code.
      operationId: join_game_games_join_post
      parameters:
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinGameRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinGameResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}:
    get:
      tags:
      - games
      summary: Get Game State
      description: Fetch current game state.
      operationId: get_game_state_games__game_id__get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/inventory:
    get:
      tags:
      - games
      summary: Get Inventory
      description: "Current question inventory \u2014 all slots grouped by type."
      operationId: get_inventory_games__game_id__inventory_get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/players/{player_id}:
    patch:
      tags:
      - games
      summary: Patch Player
      description: Update a player's role, name, or color.
      operationId: patch_player_games__game_id__players__player_id__patch
      parameters:
      - name: player_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Player Id
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/start:
    post:
      tags:
      - games
      summary: Start Game
      description: Transition the game from lobby to hiding.
      operationId: start_game_games__game_id__start_post
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/end:
    post:
      tags:
      - games
      summary: End Game
      description: Transition the game to finished.
      operationId: end_game_games__game_id__end_post
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/hider-station:
    get:
      tags:
      - games
      summary: Get Hider Station
      description: The hider's station and election status. Available during hiding
        and seeking.
      operationId: get_hider_station_games__game_id__hider_station_get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HiderStationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - games
      summary: Elect Hider Station
      description: Elect a station as the hider's hiding zone anchor. Permanent.
      operationId: elect_hider_station_games__game_id__hider_station_post
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElectStationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HidingZoneResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/nearby-stations:
    get:
      tags:
      - games
      summary: Get Nearby Stations
      description: Playable stops within hiding zone radius of a point, with hiding
        zone polygons.
      operationId: get_nearby_stations_games__game_id__nearby_stations_get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: lat
        in: query
        required: true
        schema:
          type: number
          description: Latitude of the query point.
          title: Lat
        description: Latitude of the query point.
      - name: lng
        in: query
        required: true
        schema:
          type: number
          description: Longitude of the query point.
          title: Lng
        description: Longitude of the query point.
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NearbyStationResponse'
                title: Response Get Nearby Stations Games  Game Id  Nearby Stations
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/hiding-zone:
    get:
      tags:
      - games
      summary: Get Hiding Zone
      description: Hiding zone polygon for a given station.
      operationId: get_hiding_zone_games__game_id__hiding_zone_get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: station_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          description: Stop ID to compute the hiding zone for.
          title: Station Id
        description: Stop ID to compute the hiding zone for.
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HidingZoneResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/map:
    get:
      tags:
      - games
      summary: Get Effective Map
      description: Effective map with transit data and exclusions applied.
      operationId: get_effective_map_games__game_id__map_get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectiveMapResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /games/{game_id}/candidate-stations:
    get:
      tags:
      - games
      summary: List Candidate Stations
      description: Playable stops whose hiding zone circle is not fully covered by
        exclusion zones.
      operationId: list_candidate_stations_games__game_id__candidate_stations_get
      parameters:
      - name: game_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Game Id
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset.
          default: 0
          title: Offset
        description: Pagination offset.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Pagination limit.
          default: 50
          title: Limit
        description: Pagination limit.
      - name: x-client-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Client-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StopResponse'
                title: Response List Candidate Stations Games  Game Id  Candidate
                  Stations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateGameRequest:
      properties:
        map_id:
          type: string
          format: uuid
          title: Map Id
          description: ID of the map to play on.
        excluded_stop_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Excluded Stop Ids
          description: Stop IDs to exclude from play for this game.
        excluded_route_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Excluded Route Ids
          description: Route IDs to exclude from play for this game.
        device_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Token
          description: Hex-encoded APNS device token. Optional for game creation.
        device_token_environment:
          type: string
          title: Device Token Environment
          description: 'APNS environment: "production" or "sandbox".'
          default: production
        hiding_time_min:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Hiding Time Min
          description: Override hiding phase duration. Defaults to map size.
        base_question_delay_min:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Base Question Delay Min
          description: Override auto-answer delay. Defaults to 5 minutes.
      type: object
      required:
      - map_id
      title: CreateGameRequest
      description: Create a new game on a map.
    EffectiveMapResponse:
      properties:
        name:
          type: string
          title: Name
        size:
          $ref: '#/components/schemas/MapSize'
        boundary:
          $ref: '#/components/schemas/Polygon'
          description: GeoJSON Polygon.
        districts:
          items: {}
          type: array
          title: Districts
        district_classes:
          items: {}
          type: array
          title: District Classes
        stops:
          items:
            $ref: '#/components/schemas/StopResponse'
          type: array
          title: Stops
        routes:
          items:
            $ref: '#/components/schemas/RouteResponse'
          type: array
          title: Routes
      type: object
      required:
      - name
      - size
      - boundary
      - districts
      - district_classes
      - stops
      - routes
      title: EffectiveMapResponse
      description: 'The game map with transit data and exclusions applied.


        Single source of truth for what the client should render.'
    ElectStationRequest:
      properties:
        station_id:
          type: string
          format: uuid
          title: Station Id
          description: ID of the stop to elect as hider station.
        location:
          $ref: '#/components/schemas/Point'
          description: Current hider position as a GeoJSON Point.
      type: object
      required:
      - station_id
      - location
      title: ElectStationRequest
      description: "Elect a hider station \u2014 lock in the hiding zone."
...
    GameResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        map_id:
          type: string
          format: uuid
          title: Map Id
        status:
          $ref: '#/components/schemas/GameStatus'
        convention:
          type: string
          title: Convention
          description: 'Distance convention: "metric" or "imperial".'
        join_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Join Code
          description: 4-character code for joining. Null once hiding starts.
        hiding_time_min:
          type: integer
          title: Hiding Time Min
          description: Hiding phase duration in minutes.
        base_question_delay_min:
          type: integer
          title: Base Question Delay Min
          description: Auto-answer delay in minutes.
        inventory:
          $ref: '#/components/schemas/InventoryResponse'
          description: "Question inventory \u2014 all slots grouped by type."
        players:
          items:
            $ref: '#/components/schemas/PlayerResponse'
          type: array
          title: Players
        created_at:
          type: string
          format: date-time
          title: Created At
        hiding_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Hiding Started At
          description: When the hiding phase began. Clients compute countdown.
        seeking_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Seeking Started At
          description: When the seeking phase began.
      type: object
      required:
      - id
      - map_id
      - status
      - convention
      - join_code
      - hiding_time_min
      - base_question_delay_min
      - inventory
      - players
      - created_at
      title: GameResponse
      description: Full game state, including players and question inventory.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HiderStationResponse:
      properties:
        hider_station_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Hider Station Id
          description: Station UUID, or null if not yet assigned (ambiguous/pending).
        station_election_status:
          $ref: '#/components/schemas/StationElectionStatus'
          description: Current station election status.
      type: object
      required:
      - hider_station_id
      - station_election_status
      title: HiderStationResponse
      description: "The hider's assigned station \u2014 includes election status."
...
    InventoryResponse:
      properties:
        radar_slots:
          items:
            $ref: '#/components/schemas/SlotResponse'
          type: array
          title: Radar Slots
          description: Radar question slots.
        thermometer_slots:
          items:
            $ref: '#/components/schemas/SlotResponse'
          type: array
          title: Thermometer Slots
          description: Thermometer question slots.
        matching_slots:
          items:
            $ref: '#/components/schemas/SlotResponse'
          type: array
          title: Matching Slots
          description: Matching question slots.
        measuring_slots:
          items:
            $ref: '#/components/schemas/SlotResponse'
          type: array
          title: Measuring Slots
          description: Measuring question slots.
      type: object
      required:
      - radar_slots
      - thermometer_slots
      - matching_slots
      - measuring_slots
      title: InventoryResponse
      description: "Question inventory \u2014 all slots grouped by question type."
    JoinGameRequest:
      properties:
        join_code:
          type: string
          title: Join Code
          description: 4-character code shared by the host.
        name:
          type: string
          title: Name
          description: Display name for this player.
        color:
          type: string
          title: Color
          description: Hex color for this player, e.g. "#FF5733".
        role:
          anyOf:
          - $ref: '#/components/schemas/PlayerRole'
          - type: 'null'
          description: Optional role to assign on join (hider or seeker).
        device_token:
          type: string
          title: Device Token
          description: "Hex-encoded APNS device token. Required \u2014 push is central\
            \ to gameplay."
        device_token_environment:
          type: string
          title: Device Token Environment
          description: 'APNS environment: "production" or "sandbox".'
          default: production
      type: object
      required:
      - join_code
      - name
      - color
      - device_token
      title: JoinGameRequest
      description: Join an existing game by its join code.
    JoinGameResponse:
      properties:
        game:
          $ref: '#/components/schemas/GameResponse'
        player_id:
          type: string
          format: uuid
          title: Player Id
          description: The joining player's ID for subsequent requests.
      type: object
      required:
      - game
      - player_id
      title: JoinGameResponse
      description: "Returned when a player joins a game \u2014 includes the game state\
        \ and the caller's player ID."
...
    PlayerResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        color:
          type: string
          title: Color
          description: Hex color, e.g. "#FF5733".
        role:
          anyOf:
          - $ref: '#/components/schemas/PlayerRole'
          - type: 'null'
          description: Null until the host assigns a role.
      type: object
      required:
      - id
      - name
      - color
      - role
      title: PlayerResponse
      description: A player in a game.
...
    PlayerUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: New display name.
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
          description: New hex color.
        role:
          anyOf:
          - $ref: '#/components/schemas/PlayerRole'
          - type: 'null'
          description: Assign hider or seeker role.
      type: object
      title: PlayerUpdate
      description: Partial update to a player. All fields are optional; only provided
        fields are applied.
...